ai-service/requirements.txt
flask
flask-cors
opencv-python
numpy
tensorflow
keras
scikit-learn
pandas
pillow
python-dotenv
requests
